refactor: move detection of RdbEngine by JDBC url from JdbcConfig to RdbEngineStrategy
