+++ core/src/main/java/com/scalar/db/storage/jdbc/RdbEngineSqlServer.java
@@ -50,7 +50,7 @@ class RdbEngineSqlServer extends RdbEngineStrategy {
   }
 
   @Override
-  protected RdbEngine getRdbEngine() {
+  public RdbEngine getRdbEngine() {
     return RdbEngine.SQL_SERVER;
   }
