+++ core/src/main/java/com/scalar/db/storage/jdbc/RdbEngineStrategy.java
@@ -24,7 +24,7 @@ public abstract class RdbEngineStrategy {
   public abstract RdbEngineErrorType interpretSqlException(SQLException e);
 
   public static RdbEngineStrategy create(JdbcConfig config) {
-    return create(config.getRdbEngine());
+    return create(config.getJdbcUrl());
   }
 
   public static RdbEngineStrategy create(RdbEngine rdbEngine) {
@@ -43,7 +43,21 @@ public abstract class RdbEngineStrategy {
     }
   }
 
-  protected abstract RdbEngine getRdbEngine();
+  static RdbEngineStrategy create(String jdbcUrl) {
+    if (jdbcUrl.startsWith("jdbc:mysql:")) {
+      return new RdbEngineMysql();
+    } else if (jdbcUrl.startsWith("jdbc:postgresql:")) {
+      return new RdbEnginePostgresql();
+    } else if (jdbcUrl.startsWith("jdbc:oracle:")) {
+      return new RdbEngineOracle();
+    } else if (jdbcUrl.startsWith("jdbc:sqlserver:")) {
+      return new RdbEngineSqlServer();
+    } else {
+      throw new IllegalArgumentException("the rdb engine is not supported: " + jdbcUrl);
+    }
+  }
+
+  public abstract RdbEngine getRdbEngine();
 
   protected abstract String getDataTypeForEngine(DataType dataType);
