+++ core/src/main/java/com/scalar/db/transaction/jdbc/JdbcTransactionManager.java
@@ -11,11 +11,7 @@ import com.scalar.db.common.TableMetadataManager;
 import com.scalar.db.common.checker.OperationChecker;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.transaction.TransactionException;
-import com.scalar.db.storage.jdbc.JdbcAdmin;
-import com.scalar.db.storage.jdbc.JdbcConfig;
-import com.scalar.db.storage.jdbc.JdbcService;
-import com.scalar.db.storage.jdbc.JdbcUtils;
-import com.scalar.db.storage.jdbc.RdbEngine;
+import com.scalar.db.storage.jdbc.*;
 import com.scalar.db.storage.jdbc.query.QueryBuilder;
 import java.sql.SQLException;
 import java.util.UUID;
@@ -30,7 +26,7 @@ public class JdbcTransactionManager extends ActiveTransactionManagedDistributedT
 
   private final BasicDataSource dataSource;
   private final BasicDataSource tableMetadataDataSource;
-  private final RdbEngine rdbEngine;
+  private final RdbEngineStrategy rdbEngine;
   private final JdbcService jdbcService;
 
   @Inject
@@ -39,7 +35,7 @@ public class JdbcTransactionManager extends ActiveTransactionManagedDistributedT
     JdbcConfig config = new JdbcConfig(databaseConfig);
 
     dataSource = JdbcUtils.initDataSource(config, true);
-    rdbEngine = config.getRdbEngine();
+    rdbEngine = RdbEngineStrategy.create(config);
 
     tableMetadataDataSource = JdbcUtils.initDataSourceForTableMetadata(config);
     TableMetadataManager tableMetadataManager =
@@ -48,7 +44,7 @@ public class JdbcTransactionManager extends ActiveTransactionManagedDistributedT
             databaseConfig.getMetadataCacheExpirationTimeSecs());
 
     OperationChecker operationChecker = new OperationChecker(tableMetadataManager);
-    QueryBuilder queryBuilder = new QueryBuilder(rdbEngine);
+    QueryBuilder queryBuilder = new QueryBuilder(rdbEngine.getRdbEngine());
     jdbcService = new JdbcService(tableMetadataManager, operationChecker, queryBuilder);
   }
 
@@ -62,7 +58,7 @@ public class JdbcTransactionManager extends ActiveTransactionManagedDistributedT
     super(databaseConfig);
     this.dataSource = dataSource;
     this.tableMetadataDataSource = tableMetadataDataSource;
-    this.rdbEngine = rdbEngine;
+    this.rdbEngine = RdbEngineStrategy.create(rdbEngine);
     this.jdbcService = jdbcService;
   }
 
@@ -76,7 +72,7 @@ public class JdbcTransactionManager extends ActiveTransactionManagedDistributedT
   public DistributedTransaction begin(String txId) throws TransactionException {
     try {
       JdbcTransaction transaction =
-          new JdbcTransaction(txId, jdbcService, dataSource.getConnection(), rdbEngine);
+          new JdbcTransaction(txId, jdbcService, dataSource.getConnection(), rdbEngine.getRdbEngine());
       getNamespace().ifPresent(transaction::withNamespace);
       getTable().ifPresent(transaction::withTable);
       return decorate(transaction);
