+++ core/src/test/java/com/scalar/db/storage/jdbc/JdbcAdminTestBase.java
@@ -5,11 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 import com.scalar.db.api.Scan.Ordering.Order;
 import com.scalar.db.api.TableMetadata;
@@ -31,9 +27,7 @@ import java.util.Set;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+import org.mockito.*;
 import org.mockito.invocation.InvocationOnMock;
 
 /**
@@ -61,9 +55,12 @@ public abstract class JdbcAdminTestBase {
 
   private JdbcAdmin createJdbcAdminFor(RdbEngine rdbEngine) {
     // Arrange
-    when(config.getRdbEngine()).thenReturn(rdbEngine);
-
-    return new JdbcAdmin(dataSource, config);
+    RdbEngineStrategy st = RdbEngineStrategy.create(rdbEngine);
+    try(MockedStatic<RdbEngineStrategy> mocked = mockStatic(RdbEngineStrategy.class)) {
+      mocked.when(() -> RdbEngineStrategy.create(any(JdbcConfig.class)))
+          .thenReturn(st);
+      return new JdbcAdmin(dataSource, config);
+    }
   }
 
   /**
